fix(enc): tests running in parallel bug
