+++ config/default.js
@@ -53,7 +53,7 @@ module.exports = {
   NODE_HASHLENGTH: 32,
   ZERO: '0x0000000000000000000000000000000000000000000000000000000000000000',
   HASH_TYPE: 'mimc',
-  USE_STUBS: process.env.USE_STUBS === 'true',
+  USE_STUBS: process.env.USE_STUBS === 'false',
   VK_IDS: { deposit: 0, single_transfer: 1, double_transfer: 2, withdraw: 3 }, // used as an enum to mirror the Shield contracts enum for vk types. The keys of this object must correspond to a 'folderpath' (the .zok file without the '.zok' bit)
 
   // the various parameters needed to describe the Babyjubjub curve that we use for El-Gamal
