+++ docker-compose.yml
@@ -2,7 +2,7 @@ version: '3.5'
 # Use this script for running up nightfall_2 in 'developer' mode with local
 # bindings.  See the readme for more information.
 services:
-  client:
+  client1:
     build:
       dockerfile: Dockerfile
       context: ./nightfall-client
@@ -10,9 +10,9 @@ services:
       - ./nightfall-client/src:/app/src
       - ./config/default.js:/app/config/default.js
       - build:/app/build
-      - mongodb:/app/mongodb
+      - mongodb1:/app/mongodb
     networks:
-      - nightfall_network
+      - nightfall_network1
     ports:
       - 27017:27017
       - 8080:80
@@ -20,8 +20,8 @@ services:
       - deployer
       - worker
       - ganache
-      - timber
-      - rabbitmq
+      - timber1
+      - rabbitmq1
     environment:
       MONGO_INITDB_ROOT_USERNAME: admin
       MONGO_INITDB_ROOT_PASSWORD: admin
@@ -29,9 +29,50 @@ services:
       BLOCKCHAIN_WS_HOST: ganache
       BLOCKCHAIN_PORT: 8545
       ZOKRATES_WORKER_HOST: worker
-      RABBITMQ_HOST: amqp://rabbitmq
+      RABBITMQ_HOST: amqp://rabbitmq1
       RABBITMQ_PORT: 5672
+      TIMBER_HOST: timber1
+      TIMBER_PORT: 80
+      OPTIMIST_HOST: optimist1
+      OPTIMIST_PORT: 80
       ENABLE_QUEUE: 1
+      USE_STUBS: 'true' # make sure this flag is the same as in deployer service
+    command: ['npm', 'run', 'dev']
+
+  client2:
+    build:
+      dockerfile: Dockerfile
+      context: ./nightfall-client
+    volumes:
+      - ./nightfall-client/src:/app/src
+      - ./config/default.js:/app/config/default.js
+      - build:/app/build
+      - mongodb2:/app/mongodb
+    networks:
+      - nightfall_network2
+    ports:
+      - 27018:27017
+      - 8084:80
+    depends_on:
+      - deployer
+      - worker
+      - ganache
+      - timber2
+      - rabbitmq2
+    environment:
+      MONGO_INITDB_ROOT_USERNAME: admin
+      MONGO_INITDB_ROOT_PASSWORD: admin
+      LOG_LEVEL: debug
+      BLOCKCHAIN_WS_HOST: ganache
+      BLOCKCHAIN_PORT: 8545
+      ZOKRATES_WORKER_HOST: worker
+      RABBITMQ_HOST: amqp://rabbitmq2
+      RABBITMQ_PORT: 5672
+      TIMBER_HOST: timber2
+      TIMBER_PORT: 80
+      ENABLE_QUEUE: 1
+      OPTIMIST_HOST: optimist2
+      OPTIMIST_PORT: 80
       USE_STUBS: 'true' # make sure this flag is the same as in deployer service
     command: ['npm', 'run', 'dev']
 
@@ -56,7 +97,8 @@ services:
     depends_on:
       - deployer
     networks:
-      - nightfall_network
+      - nightfall_network1
+      - nightfall_network2
     environment:
       LOG_LEVEL: info
 
@@ -88,7 +130,8 @@ services:
         source: ./nightfall-deployer/migrations
         target: /app/migrations
     networks:
-      - nightfall_network
+      - nightfall_network1
+      - nightfall_network2
     environment:
       LOG_LEVEL: debug
       # ETH_NETWORK sets the network selected by Truffle from truffle-config.js
@@ -96,6 +139,8 @@ services:
       BLOCKCHAIN_WS_HOST: ganache
       BLOCKCHAIN_PORT: 8545
       ZOKRATES_WORKER_HOST: worker
+      # TIMBER_HOST: timber1
+      # TIMBER_PORT: 80
       USE_STUBS: 'true'
 
   # The Ganache client, containing required precompiles
@@ -106,10 +151,11 @@ services:
     ports:
       - '8545:8545'
     networks:
-      - nightfall_network
+      - nightfall_network1
+      - nightfall_network2
 
   # Timber service, configured for nightfall
-  timber:
+  timber1:
     build:
       dockerfile: Dockerfile
       context: ./timber
@@ -117,9 +163,9 @@ services:
     depends_on:
       - ganache
       - deployer
-      - timber-database
+      - timber-database1
     networks:
-      - nightfall_network
+      - nightfall_network1
     ports:
       - 8083:80
     volumes:
@@ -135,7 +181,41 @@ services:
     environment:
       BLOCKCHAIN_HOST: ws://ganache
       BLOCKCHAIN_PORT: 8545
-      MONGO_HOST: timber-database
+      MONGO_HOST: timber-database1
+      MONGO_PORT: 27017
+      MONGO_NAME: merkle_tree
+      HASH_TYPE: mimc
+      LOG_LEVEL: info
+      AUTOSTART: enabled
+
+  # Timber service, configured for nightfall
+  timber2:
+    build:
+      dockerfile: Dockerfile
+      context: ./timber
+    restart: on-failure
+    depends_on:
+      - ganache
+      - deployer
+      - timber-database2
+    networks:
+      - nightfall_network2
+    ports:
+      - 8087:80
+    volumes:
+      - type: volume
+        source: build
+        target: /app/build/
+      - type: bind
+        source: ./timber/src
+        target: /app/src/
+      - type: bind
+        source: ./config/timber.js
+        target: /app/config/default.js
+    environment:
+      BLOCKCHAIN_HOST: ws://ganache
+      BLOCKCHAIN_PORT: 8545
+      MONGO_HOST: timber-database2
       MONGO_PORT: 27017
       MONGO_NAME: merkle_tree
       HASH_TYPE: mimc
@@ -143,32 +223,50 @@ services:
       AUTOSTART: enabled
 
   #The database storing the merkle tree
-  timber-database:
+  timber-database1:
     image: mongo
     environment:
       - MONGO_INITDB_DATABASE=merkle_tree
     volumes:
-      - timber-database-volume:/data/db
+      - timber-database-volume1:/data/db
     networks:
-      - nightfall_network
+      - nightfall_network1
 
-  rabbitmq:
+  #The database storing the merkle tree
+  timber-database2:
+    image: mongo
+    environment:
+      - MONGO_INITDB_DATABASE=merkle_tree
+    volumes:
+      - timber-database-volume2:/data/db
+    networks:
+      - nightfall_network2
+
+  rabbitmq1:
     image: rabbitmq
     ports:
       - '15674:15674'
       - '5672:5672'
     networks:
-      - nightfall_network
+      - nightfall_network1
 
-  optimist:
+  rabbitmq2:
+    image: rabbitmq
+    ports:
+      - '15675:15674'
+      - '5673:5672'
+    networks:
+      - nightfall_network2
+
+  optimist1:
     build:
       context: ./nightfall-optimist
     depends_on:
       - ganache
-      - timber
+      - timber1
       - deployer
     networks:
-      - nightfall_network
+      - nightfall_network1
     ports:
       - 8081:80
       # websocket port for Optimist is on localhost:8082
@@ -187,13 +285,49 @@ services:
       LOG_LEVEL: debug
       IS_CHALLENGER: 'true'
       TRANSACTIONS_PER_BLOCK: 2
+      TIMBER_HOST: timber1
+      TIMBER_PORT: 80
+    command: ['npm', 'run', 'dev']
+
+  optimist2:
+    build:
+      context: ./nightfall-optimist
+    depends_on:
+      - ganache
+      - timber2
+      - deployer
+    networks:
+      - nightfall_network2
+    ports:
+      - 8085:80
+      # websocket port for Optimist is on localhost:8086
+      - 8086:8080
+    volumes:
+      - type: volume
+        source: build
+        target: /app/build/
+      - ./nightfall-optimist/src:/app/src
+      - ./config/default.js:/app/config/default.js
+    environment:
+      WEBSOCKET_PORT: 8084
+      BLOCKCHAIN_WS_HOST: ganache
+      BLOCKCHAIN_PORT: 8545
+      HASH_TYPE: mimc
+      LOG_LEVEL: debug
+      IS_CHALLENGER: 'true'
+      TRANSACTIONS_PER_BLOCK: 2
+      TIMBER_HOST: timber2
+      TIMBER_PORT: 80
     command: ['npm', 'run', 'dev']
 
 volumes:
-  mongodb:
+  mongodb1:
+  mongodb2:
   proving_files:
   build:
-  timber-database-volume:
+  timber-database-volume1:
+  timber-database-volume2:
 
 networks:
-  nightfall_network:
+  nightfall_network1:
+  nightfall_network2:
